Adding a working version of my hearthstone project
