+++ Journal.md
@@ -34,7 +34,14 @@ Tasks completed:
 I implemented this week's hibernate activity into my project and modified my database to reflect that new technology. I also put some serious thought into how I how I want my database to perform and the columns fields and areas that I want it to reflect. 
 
 
+### Week 6
 
+Tasks completed:
+* Almost finished setting up the AWS. Running into a problem with the database being able to connect with the hibernate driver.
+* I destroyed my project, and then cloned from an existing version. I consider this a success because the project was in such a sorry condition. 
+* Added packages and got my log file working again. 
+
+I spent much of my time this week reviewing my peers' code. I also set up my AWS instance and have the ability to ssh into it. Next steps will include thinking about how the controllers for my classes are going to work and implementing servlets for those classes.
