+++ src/test/java/com/hearthstone/persistence/DeckDaoTest.java
@@ -1,10 +1,9 @@
-package persistence;
+package com.hearthstone.persistence;
 
-import entity.Deck;
-import entity.Stats;
+import com.hearthstone.entity.Deck;
+import com.hearthstone.entity.Stats;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import test.util.Database;
 
 import static org.junit.jupiter.api.Assertions.*;
