+++ src/test/resources/hibernate.cfg.xml
@@ -17,8 +17,8 @@
         <property name="hibernate.c3p0.idle_test_period">3000</property>
 
 
-        <mapping class="entity.Deck"/>
-        <mapping class="entity.Stats"/>
+        <mapping class="com.hearthstone.entity.Deck"/>
+        <mapping class="com.hearthstone.entity.Stats"/>
 
 
     </session-factory>
