+++ src/test/resources/log4j2.properties
@@ -1,5 +1,5 @@
 
-property.filepath = /home/student/IdeaProjects/indieProjectHerrmann.logs
+property.filepath = /home/student/IdeaProjects/hearthstoneDeckTracker1/logs
 appenders = console, file, hibernate
 
 appender.console.type = Console
@@ -34,7 +34,7 @@ appender.hibernate.strategy.delete.ifAccumulatedFileCount.exceeds = 5
 
 loggers=file, hibernate
 
-logger.file.name = edu.matc
+logger.file.name = com.hearthstone
 logger.file.level = debug
 logger.file.appenderRefs = file
 logger.file.appenderRef.file.ref = LOGFILE
